requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.11.1
ffmpeg-python==0.2.0
python-dotenv==1.1.1
python-multipart==0.0.9
yt-dlp==2024.8.6
numpy==1.26.4
torch==2.2.2
regex==2024.11.6
tiktoken==0.11.0
llvmlite==0.44.0
numba==0.61.2
openai-whisper==20231117
asyncpg==0.29.0
pytest==8.4.2
pytest-asyncio==1.2.0
huggingface_hub==0.35.1
gradio==4.44.1
soundfile==0.12.1

